[Feature] Add some view methods.

+++ src/AElf.Kernel.Consensus.AEDPoS/ConsensusOptions.cs
@@ -11,6 +11,6 @@ namespace AElf.Kernel.Consensus.AEDPoS
         public int MiningInterval { get; set; }
         public DateTime StartTimestamp { get; set; } = DateTime.MinValue;
         public long InitialTermNumber { get; set; }
-        public int TimeEachTerm { get; set; }
+        public long TimeEachTerm { get; set; }
     }
 }
